#2133 refactored code
